+++ src/pymmcore_widgets/_stack_viewer_v2/__init__.py
