+++ src/pymmcore_widgets/_stack_viewer_v2/_backends/_vispy.py
