+++ src/pymmcore_widgets/_stack_viewer_v2/_dims_slider.py
