+++ src/pymmcore_widgets/_stack_viewer_v2/_lut_control.py
