+++ src/pymmcore_widgets/experimental.py
@@ -1,4 +1,4 @@
-from ._stack_viewer import MDAViewer
 from ._stack_viewer_v1 import StackViewer
+from ._stack_viewer_v2 import MDAViewer
 
 __all__ = ["StackViewer", "MDAViewer"]
